+++ src/Core/Dfe.Complete.Domain/Entities/Project.cs
@@ -428,6 +428,7 @@ public class Project : BaseAggregateRoot, IEntity<ProjectId>
             {
                 Id = new NoteId(Guid.NewGuid()),
                 CreatedAt = note.CreatedAt,
+                UpdatedAt = note.UpdatedAt,
                 Body = note.Body,
                 ProjectId = note.ProjectId,
                 TaskIdentifier = note.TaskIdentifier,
@@ -470,5 +471,24 @@ public class Project : BaseAggregateRoot, IEntity<ProjectId>
     public void AddAcademyUrn(Urn urn)
     {
         AcademyUrn = urn;
+    }
+    
+    public void UpdateSignificantDate(DateOnly date)
+    {
+        SignificantDate = date;
+    }
+    
+    public void AddSignificantDateHistory(SignificantDateHistory significantDateHistory)
+    {
+        SignificantDateHistories.Add(new SignificantDateHistory
+        {
+            Id = significantDateHistory.Id,
+            CreatedAt = significantDateHistory.CreatedAt,
+            UpdatedAt = significantDateHistory.UpdatedAt,
+            ProjectId = significantDateHistory.ProjectId,
+            UserId = significantDateHistory.UserId,
+            PreviousDate = significantDateHistory.PreviousDate,
+            RevisedDate = significantDateHistory.RevisedDate,
+        });
     }
 }