+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Customizations/CustomWebApplicationDbContextFactoryCustomization.cs
@@ -6,7 +6,7 @@ using Dfe.AcademiesApi.Client.Contracts;
 using Dfe.AcademiesApi.Client.Security;
 using Dfe.AcademiesApi.Client.Settings;
 using Dfe.Complete.Api.Client.Extensions;
-using Dfe.Complete.Application.Common.Mappers;
+using Dfe.Complete.Application.Mappers;
 using Dfe.Complete.Client;
 using Dfe.Complete.Client.Contracts;
 using Dfe.Complete.Infrastructure.Database;
@@ -116,6 +116,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
                 services.AddCompleteApiClient<ICsvExportClient, CsvExportClient>(config, client);
                 services.AddCompleteApiClient<IUsersClient, UsersClient>(config, client);
                 services.AddCompleteApiClient<IServiceSupportClient, ServiceSupportClient>(config, client);
+                services.AddCompleteApiClient<IContactsClient, ContactsClient>(config, client);
                 var serviceProvider = services.BuildServiceProvider();
 
                 fixture.Inject(factory);
@@ -127,6 +128,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
                 fixture.Inject(serviceProvider.GetRequiredService<ICsvExportClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IUsersClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IServiceSupportClient>());
+                fixture.Inject(serviceProvider.GetRequiredService<IContactsClient>());
                 fixture.Inject(new List<Claim>());
 
                 return factory;