+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/projects/all-projects.cy.ts
@@ -34,6 +34,7 @@ const transferSchoolName = "Abbey College Manchester";
 const transferRegion = "North West";
 const transferFormAMatProject = ProjectBuilder.createTransferFormAMatProjectRequest();
 const transferFormAMatSchoolName = "Priory Rise School";
+const nextMonthShortUS = `${nextMonth.toLocaleString("en-US", { month: "short" })} ${nextMonth.getFullYear()}`; // bug 228624
 
 describe("View all projects", () => {
     before(() => {
@@ -189,8 +190,8 @@ describe("View all projects", () => {
             .columnHasValue("Local authority", localAuthorityShort)
             .columnHasValue("Incoming trust", macclesfieldTrust.name.toUpperCase()) // bug 208086
             .columnHasValue("All conditions met", "Not yet")
-            .columnHasValue("Confirmed date (Original date)", nextMonthShort)
-            .goTo(`${schoolName} ${project.urn.value}`);
+            .columnHasValue("Confirmed date (Original date)", nextMonthShortUS) // bug 228624
+            .goTo(schoolName);
         projectDetailsPage.containsHeading(schoolName);
     });
 
@@ -213,7 +214,7 @@ describe("View all projects", () => {
             .columnHasValue("Outgoing trust", macclesfieldTrust.name.toUpperCase()) // bug 208086
             .columnHasValue("Incoming trust", dimensionsTrust.name.toUpperCase()) // bug 208086
             .columnHasValue("Authority to proceed", "Not yet")
-            .columnHasValue("Confirmed date (Original date)", nextMonthShort)
+            .columnHasValue("Confirmed date (Original date)", nextMonthShortUS) // bug 228624
             .goTo(`${transferSchoolName} ${transferProject.urn.value}`);
         projectDetailsPage.containsHeading(transferSchoolName);
     });